code reorganisation, cleanup, small optimisations
